Insert new conta no BD, alterações no layout de ContaInfo e novo icone do app

+++ app/src/main/java/br/edu/ifsp/controlefinancas/activity/model/Conta.java
@@ -9,7 +9,7 @@ public class Conta {
     public Conta() {
     }
 
-    public Conta(long id, String descricao, float saldo) {
+    public Conta(long id, String descricao, double saldo) {
         this.id = id;
         this.descricao = descricao;
         this.saldo = saldo;
